src/main/webapp/app/entities/raw-asset-proc-tmp/raw-asset-proc-tmp.component.ts
import { defineComponent, inject, onMounted, ref, type Ref, watch } from 'vue';
import { useI18n } from 'vue-i18n';

import RawAssetProcTmpService from './raw-asset-proc-tmp.service';
import { type IRawAssetProcTmp } from '@/shared/model/raw-asset-proc-tmp.model';
import useDataUtils from '@/shared/data/data-utils.service';
import { useAlertService } from '@/shared/alert/alert.service';

export default defineComponent({
  compatConfig: { MODE: 3 },
  name: 'RawAssetProcTmp',
  setup() {
    const { t: t$ } = useI18n();
    const dataUtils = useDataUtils();
    const rawAssetProcTmpService = inject('rawAssetProcTmpService', () => new RawAssetProcTmpService());
    const alertService = inject('alertService', () => useAlertService(), true);

    const itemsPerPage = ref(20);
    const queryCount: Ref<number> = ref(null);
    const page: Ref<number> = ref(1);
    const propOrder = ref('id');
    const reverse = ref(false);
    const totalItems = ref(0);

    const rawAssetProcTmps: Ref<IRawAssetProcTmp[]> = ref([]);

    const isFetching = ref(false);

    const clear = () => {
      page.value = 1;
    };

    const sort = (): Array<any> => {
      const result = [propOrder.value + ',' + (reverse.value ? 'desc' : 'asc')];
      if (propOrder.value !== 'id') {
        result.push('id');
      }
      return result;
    };

    const retrieveRawAssetProcTmps = async () => {
      isFetching.value = true;
      try {
        const paginationQuery = {
          page: page.value - 1,
          size: itemsPerPage.value,
          sort: sort(),
        };
        const res = await rawAssetProcTmpService().retrieve(paginationQuery);
        totalItems.value = Number(res.headers['x-total-count']);
        queryCount.value = totalItems.value;
        rawAssetProcTmps.value = res.data;
      } catch (err) {
        alertService.showHttpError(err.response);
      } finally {
        isFetching.value = false;
      }
    };

    const handleSyncList = () => {
      retrieveRawAssetProcTmps();
    };

    onMounted(async () => {
      await retrieveRawAssetProcTmps();
    });

    const removeId: Ref<number> = ref(null);
    const removeEntity = ref<any>(null);
    const prepareRemove = (instance: IRawAssetProcTmp) => {
      removeId.value = instance.id;
      removeEntity.value.show();
    };
    const closeDialog = () => {
      removeEntity.value.hide();
    };
    const removeRawAssetProcTmp = async () => {
      try {
        await rawAssetProcTmpService().delete(removeId.value);
        const message = t$('azimuteErpSpringVueMonolith04App.rawAssetProcTmp.deleted', { param: removeId.value }).toString();
        alertService.showInfo(message, { variant: 'danger' });
        removeId.value = null;
        retrieveRawAssetProcTmps();
        closeDialog();
      } catch (error) {
        alertService.showHttpError(error.response);
      }
    };

    const changeOrder = (newOrder: string) => {
      if (propOrder.value === newOrder) {
        reverse.value = !reverse.value;
      } else {
        reverse.value = false;
      }
      propOrder.value = newOrder;
    };

    // Whenever order changes, reset the pagination
    watch([propOrder, reverse], async () => {
      if (page.value === 1) {
        // first page, retrieve new data
        await retrieveRawAssetProcTmps();
      } else {
        // reset the pagination
        clear();
      }
    });

    // Whenever page changes, switch to the new page.
    watch(page, async () => {
      await retrieveRawAssetProcTmps();
    });

    return {
      rawAssetProcTmps,
      handleSyncList,
      isFetching,
      retrieveRawAssetProcTmps,
      clear,
      removeId,
      removeEntity,
      prepareRemove,
      closeDialog,
      removeRawAssetProcTmp,
      itemsPerPage,
      queryCount,
      page,
      propOrder,
      reverse,
      totalItems,
      changeOrder,
      t$,
      ...dataUtils,
    };
  },
});
